requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
PyJWT
redis
pwdlib[argon2]
jinja2
psycopg2-binary
email-validator
python-multipart
gunicorn
alembic
tiktoken
python-magic
aioboto3
pydantic-settings
pdfplumber
python-docx
beautifulsoup4
python-pptx
odfpy
striprtf
upstash‑vector